tests(helpers): separate dns_mock module
